* local_py3:
  first changes for ckan 2.9
  Add `future` as dependency
  Python 3 support
  Add CKAN 2.9 support
  Update requirements for master's default to py3

# Conflicts:
#	ckanext/hierarchy/templates/organization/index.html

+++ bin/travis-build.bash
@@ -33,7 +33,12 @@ then
 fi
 
 python setup.py develop
-pip install -r requirements.txt
+if [ -f requirements-py2.txt ]
+then
+    pip install -r requirements-py2.txt
+else
+    pip install -r requirements.txt
+fi
 pip install -r dev-requirements.txt
 cd -
 
@@ -57,6 +62,7 @@ cd -
 echo "Installing ckanext-hierarchy and its requirements..."
 python setup.py develop
 pip install -r dev-requirements.txt
+pip install -r requirements.txt
 
 echo "Moving test.ini into a subdir..."
 mkdir subdir
